Extract `CoreData` managed model classes to separated files
